CHROMIUM: Input: atmel_mxt_ts - make size/instances u16
Previously size and instances fields were u8 but could hold up to a value
of 255.  Since we now +1 to size and instances, they must be u16 to hold
the full size of up to 256.


BUG=chromium-os:27713
TEST=Confirm object table is read correctly:
     cat /sys/bus/i2c/devices/<dev>/object

Reviewed-on: https://gerrit.chromium.org/gerrit/18548
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>

+++ drivers/input/touchscreen/atmel_mxt_ts.c
@@ -236,8 +236,8 @@ struct mxt_info {
 struct mxt_object {
 	u8 type;
 	u16 start_address;
-	u8 size;
-	u8 instances;
+	u16 size;
+	u16 instances;
 	u8 num_report_ids;
 };
